+++ debug/example/activity/activity.js
@@ -1,4 +1,3 @@
-import UrlManager from "../../flexvue/core/UrlManager.class.js";
 import { ElementButton } from "../../flexvue/components/ElementsManager.class.js";
 // 이전 화면 패널 위치
 let pre_viewpage = null;
@@ -11,16 +10,18 @@ const onReady = () => {
       Activity.onClose(pre_viewpage === null || pre_viewpage === void 0 ? void 0 : pre_viewpage.id);
     }
   });
-  // #docs/start : hash 경로가 바뀔때 마다 호출 됩니다
-  const urlManager = new UrlManager(document.location.toString());
+  // 패널 미리 생성하기
+  Activity.createPanels("view1", "view2", "view3");
+  // // #docs/start : hash 경로가 바뀔때 마다 호출 됩니다
+  // const urlManager = new UrlManager(document.location.toString());
   new ElementButton('#go_right').addEventListener('click', function () {
-    history.pushState('#slideR2L', 'right', `?a=b`);
-    const activityId = Activity.onStart('edit', 'fvueSlideFromRight top-[10%]', 'fvueSlideToLeft');
+    history.pushState('#view1', 'right', `?a=b`);
+    const activityId = Activity.onStart('view1', 'fvueSlideFromRight top-[10%]', 'fvueSlideToLeft');
     Activity.setStateHistory(activityId);
   });
   new ElementButton('#go_bottom').addEventListener('click', function () {
-    history.pushState('#slideB2T', 'bottom', `?b=c`);
-    const activityId = Activity.onStart('edit2', 'fvueSlideFromBottom top-[140%]', 'fvueSlideToTop');
+    history.pushState('#view2', 'bottom', `?b=c`);
+    const activityId = Activity.onStart('view2', 'fvueSlideFromBottom top-[140%]', 'fvueSlideToTop');
     Activity.setStateHistory(activityId);
   });
 };